requirements.txt
dj-database-url==0.5.0
django>=2.2.10
django-anymail==7.0.0
django-filter==2.2.0
djangorestframework==3.10.3
docutils==0.14
Pillow==6.2.1
whitenoise==4.1.4
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2017.3
qrcode==6.1
django-friendship==1.8.1
dropbox==9.4.0
django-storages==1.7.2